Fixed up a couple of syntax errors that don't work in 3.2
These syntax errors were not picked up by 2to3-3.2 like
lambda a, (b, c): a.replace(b, c) which works in Python 2 but not
Python 3 and
try:
	False
except:
	False = 0
	True = 1
which is totally unnecessary these days since the days that checking for
that was relavant have gone.  Python 3 does not have to check for the
existance of False.  I fixed up the former and deleted the later.